[nrf fromtree] bootutil: Split header with encryption functions
Split definitions to crypto backend specific headers.

(cherry picked from commit 5a161e4cc1cd5329c073d996f0300d2661c6b768)
